remove unused component
