--- app/components/mandatenbeheer/mandaat-bestuursorganen-selector.hbs
@@ -1,35 +1,0 @@
-<div class={{if @error "ember-power-select--error"}}>
-  {{#if this.initialized}}
-    <PowerSelect
-      @loadingMessage="Aan het laden..."
-      @noMatchesMessage="Geen resultaten"
-      @searchMessage="Typ om te zoeken"
-      @searchEnabled={{true}}
-      @allowClear={{@allowClear}}
-      @disabled={{@disabled}}
-      @placeholder="Zoek mandaat"
-      @options={{this.mandaatOptions}}
-      @selected={{this.mandaat}}
-      @searchField="bestuursfunctie.label"
-      @onChange={{this.select}}
-      @triggerId="mandaat"
-      as |mandaat|
-    >
-      {{mandaat.bestuursfunctie.label}}
-    </PowerSelect>
-  {{else}}
-    <p class="skeleton-input-full au-u-margin-top"></p>
-  {{/if}}
-
-  {{#if this.alertMessage}}
-    <AuAlert
-      @title="Opgelet"
-      @skin="info"
-      @size="tiny"
-      @closable={{false}}
-      class="au-u-margin-top"
-    >
-      <p>{{this.alertMessage}}</p>
-    </AuAlert>
-  {{/if}}
-</div>
