+++ app/components/rdf-input-fields/mandataris-mandaat-selector.js
@@ -21,12 +21,15 @@ export default class MandatarisMandaatSelector extends InputFieldComponent {
   @service store;
   @service multiUriFetcher;
   @service persoonApi;
+  @service('verkiezing') verkiezingService;
 
   @tracked mandaat = null;
   @tracked isStrictBurgemeester = false;
   @tracked initialized = false;
   @tracked bestuursorganen = [];
   @tracked person;
+
+  @tracked mandaatOptions = null;
 
   constructor() {
     super(...arguments);
@@ -47,7 +50,19 @@ export default class MandatarisMandaatSelector extends InputFieldComponent {
   async load() {
     this.person = await this.findPersonInForm();
     await Promise.all([this.loadProvidedValue(), this.loadBestuursorganen()]);
+    await this.loadMandaten();
     this.initialized = true;
+  }
+
+  async loadMandaten() {
+    const mandaten = await this.store.query('mandaat', {
+      sort: 'bestuursfunctie.label',
+      include: 'bestuursfunctie',
+      'filter[bevat-in][id]': this.bestuursorganen
+        .map((o) => o.get('id'))
+        .join(','),
+    });
+    this.mandaatOptions = mandaten;
   }
 
   async loadBestuursorganen() {
@@ -89,6 +104,7 @@ export default class MandatarisMandaatSelector extends InputFieldComponent {
 
     this.mandaat = mandate;
     this.checkPersonMandates();
+    this.validatePerson();
   }
 
   async checkPersonMandates() {
@@ -114,6 +130,29 @@ export default class MandatarisMandaatSelector extends InputFieldComponent {
       resultMessage:
         'Deze persoon heeft dit mandaat al in deze bestuursperiode',
     });
+  }
+
+  async validatePerson() {
+    if (
+      this.person &&
+      this.bestuursorganen.length >= 1 &&
+      !(await this.mandaat.allowsNonElectedPersons)
+    ) {
+      const isElected = await this.verkiezingService.checkIfPersonIsElected(
+        this.person.id,
+        this.bestuursorganen.at(0)
+      );
+      if (!isElected) {
+        super.updateValidations();
+        this.warningValidations.push({
+          validationType: EXT('notElected'),
+          hasValidation: true,
+          valid: false,
+          resultMessage:
+            'De geselecteerde persoon is niet gevonden in de verkiezingslijst.',
+        });
+      }
+    }
   }
 
   async findPersonInForm() {
